+++ src/Nethermind/Nethermind.Consensus/Processing/BlockProcessor.BlockProductionTransactionPicker.cs
@@ -43,11 +43,11 @@ namespace Nethermind.Consensus.Processing
             {
                 AddingTxEventArgs args = new(transactionsInBlock.Count, currentTx, block, transactionsInBlock);
 
-                long gasRemaining = (long)block.Header.GasLimit - (long)block.GasUsed;
+                ulong gasRemaining = block.Header.GasLimit - block.GasUsed;
 
                 // No more gas available in block for any transactions,
                 // the only case we have to really stop
-                if (GasCostOf.Transaction > gasRemaining)
+                if ((ulong)GasCostOf.Transaction > gasRemaining)
                 {
                     return args.Set(TxAction.Stop, "Block full");
                 }
@@ -65,7 +65,7 @@ namespace Nethermind.Consensus.Processing
                     return args.Set(TxAction.Skip, "Null sender");
                 }
 
-                if ((long)currentTx.GasLimit > gasRemaining)
+                if (currentTx.GasLimit > gasRemaining)
                 {
                     return args.Set(TxAction.Skip, $"Not enough gas in block, gas limit {currentTx.GasLimit} > {gasRemaining}");
                 }
