+++ src/Nethermind/Nethermind.Consensus/Validators/TxValidator.cs
@@ -389,8 +389,9 @@ public sealed class GasLimitCapTxValidator : ITxValidator
     public ValidationResult IsWellFormed(Transaction transaction, IReleaseSpec releaseSpec)
     {
         long gasLimitCap = releaseSpec.GetTxGasLimitCap();
-        return transaction.GasLimit > (ulong)gasLimitCap ?
-            TxErrorMessages.TxGasLimitCapExceeded((long)transaction.GasLimit, gasLimitCap) : ValidationResult.Success;
+        return transaction.GasLimit > (ulong)gasLimitCap
+            ? TxErrorMessages.TxGasLimitCapExceeded(transaction.GasLimit, gasLimitCap)
+            : ValidationResult.Success;
     }
 }
